Picture change when before was none and delete works AGAIN...

+++ backend/controller/item.controller.js
@@ -50,7 +50,7 @@ module.exports = {
       const url = req.body.url;
       req.body.img = `${url}.jpg`;
       request(address).pipe(fs.createWriteStream(`public/img/${url}.jpg`));
-    } 
+    }
     Item.create(req.body)
       .then(item => res.send(item))
       .catch(err => res.send(err));
@@ -69,7 +69,10 @@ module.exports = {
       const urlcim = req.body.url;
       req.body.img = `${urlcim}.jpg`;
       request(address).pipe(fs.createWriteStream(`public/img/${urlcim}.jpg`));
-      deleteFile(req.body.oldImg);
+    
+      if (req.body.oldImg !== 'none') {
+        deleteFile(req.body.oldImg);
+      }
     }
     Item.findByIdAndUpdate(req.params.id, req.body, { new: true })
       .then(item => res.json(item))
